Removed Calendar.getInstance() and replaced it kotlix-datetime Clock. Now we get current day through Clock. Also change expiration check to consider time in millis instead of using days of year

+++ core_module/src/main/java/com/example/core_module/sharedpreferences/save_data/UserDataPrefHelper.kt
@@ -13,7 +13,7 @@ interface UserDataPrefHelper {
 
     fun getCountryOfUserLocation(): String?
 
-    fun getTokenDay(): Int?
+    fun getTokenDay(): Long?
 
     fun deleteToken()
 }
